# public/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Application API Documentation",
    "version": "1.0.0",
    "description": "API documentation for the generated application",
    "contact": {
      "name": "API Support"
    }
  },
  "servers": [
    {
      "url": "https://your-production-url.com",
      "description": "Production server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "apiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      }
    }
  },
  "security": [],
  "paths": {
    "/api/about": {
      "get": {
        "summary": "Get about content",
        "tags": [
          "About"
        ],
        "responses": {
          "200": {
            "description": "About content"
          },
          "404": {
            "description": "No content found"
          },
          "500": {
            "description": "Server error"
          }
        }
      },
      "put": {
        "summary": "Update about content (auth required)",
        "tags": [
          "About"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "bio": {
                    "type": "string"
                  },
                  "profile_image_url": {
                    "type": "string"
                  },
                  "email": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Content updated"
          },
          "400": {
            "description": "Validation error"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/api/auth/login": {
      "post": {
        "summary": "Login user",
        "tags": [
          "Auth"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "email",
                  "password"
                ],
                "properties": {
                  "email": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful"
          },
          "401": {
            "description": "Invalid credentials"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/api/auth/register": {
      "post": {
        "summary": "Register a new user",
        "tags": [
          "Auth"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name",
                  "email",
                  "password"
                ],
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "email": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User registered successfully"
          },
          "400": {
            "description": "Validation error"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/api/daily-status": {
      "get": {
        "summary": "Get current daily status",
        "tags": [
          "DailyStatus"
        ],
        "responses": {
          "200": {
            "description": "Daily status"
          },
          "404": {
            "description": "No status found"
          },
          "500": {
            "description": "Server error"
          }
        }
      },
      "post": {
        "summary": "Create or update daily status (auth required)",
        "tags": [
          "DailyStatus"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "date",
                  "day_number",
                  "ingredient_of_day"
                ],
                "properties": {
                  "date": {
                    "type": "string"
                  },
                  "day_number": {
                    "type": "integer"
                  },
                  "ingredient_of_day": {
                    "type": "string"
                  },
                  "weather": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Status created/updated"
          },
          "400": {
            "description": "Validation error"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/api/recipes/{id}": {
      "get": {
        "summary": "Get recipe by ID or slug",
        "tags": [
          "Recipes"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Recipe details"
          },
          "404": {
            "description": "Recipe not found"
          },
          "500": {
            "description": "Server error"
          }
        }
      },
      "put": {
        "summary": "Update recipe (auth required)",
        "tags": [
          "Recipes"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Recipe updated"
          },
          "404": {
            "description": "Recipe not found"
          },
          "500": {
            "description": "Server error"
          }
        }
      },
      "delete": {
        "summary": "Delete recipe (auth required)",
        "tags": [
          "Recipes"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Recipe deleted"
          },
          "404": {
            "description": "Recipe not found"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/api/recipes": {
      "get": {
        "summary": "Get all recipes",
        "tags": [
          "Recipes"
        ],
        "responses": {
          "200": {
            "description": "List of recipes"
          },
          "500": {
            "description": "Server error"
          }
        }
      },
      "post": {
        "summary": "Create a new recipe (auth required)",
        "tags": [
          "Recipes"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "entry_number",
                  "title",
                  "slug",
                  "date",
                  "ingredients",
                  "steps"
                ],
                "properties": {
                  "entry_number": {
                    "type": "integer"
                  },
                  "title": {
                    "type": "string"
                  },
                  "slug": {
                    "type": "string"
                  },
                  "date": {
                    "type": "string"
                  },
                  "ingredients": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  },
                  "steps": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  },
                  "hero_image_url": {
                    "type": "string"
                  },
                  "thumbnail_image_url": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Recipe created successfully"
          },
          "400": {
            "description": "Validation error"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/api/resources": {
      "get": {
        "summary": "Get all resources",
        "tags": [
          "Resources"
        ],
        "responses": {
          "200": {
            "description": "List of resources"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/api/social-links": {
      "get": {
        "summary": "Get all social links",
        "tags": [
          "SocialLinks"
        ],
        "responses": {
          "200": {
            "description": "List of social links"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    }
  },
  "tags": []
}